+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Change Log
 
 * Use measured alpha and beta offsets when transforming from tangent to positioner.
 * Replace error in ``iauPmsafe`` with warning.
+* Add ``RoughTransform`` and ``ZhaoBurgeTransform``.
 
 
 1.1.1 (2021-10-28)